cpapp_backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
python-dotenv>=1.0.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
langchain>=0.1.0
langchain_core>=0.1.11
langchain_openai>=0.0.5
langchain-community>=0.0.10
openai>=1.13.3
requests>=2.31.0
uuid>=1.30 